SelectionItemViewHolder вынесен в отдельный файл
